+++ src/refactoring/components/Admin/AdminPage.tsx
@@ -66,17 +66,13 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
             onProductUpdate = {onProductUpdate}
             products = {products}
           />
-          
+
         </div>
 
         <div>
           <h2 className="text-2xl font-semibold mb-4">쿠폰 관리</h2>
           <div className="bg-white p-4 rounded shadow">
-            <AddCouponForm
-              newCoupon = {newCoupon}
-              setNewCoupon = {setNewCoupon}
-              handleAddCoupon = {handleAddCoupon}
-            />
+            <AddCouponForm />
             <div>
               <h3 className="text-lg font-semibold mb-2">현재 쿠폰 목록</h3>
               <AdminCouponList coupons = {coupons}/>
